Break osm file into submodule
