+++ TrailPrint3D/utils/osm/__init__.py
@@ -1,0 +1,1 @@
+"""OSM utility functions for TrailPrint3D."""
